feat: simplify match_method
