+++ crates/assertion-da/da-server/src/api/process_request.rs
@@ -1,38 +1,18 @@
-use crate::api::assertion_submission::raw_bytecode_assertion;
+use crate::api::assertion_submission::accept_bytecode_assertion;
+use crate::api::assertion_submission::accept_solidity_assertion;
+use crate::api::assertion_submission::retreive_assertion;
 use std::{
     net::SocketAddr,
     sync::Arc,
     time::Instant,
 };
 
-use crate::{
-    api::{
-        source_compilation::compile_solidity,
-        types::{
-            DbOperation,
-            DbRequest,
-            DbRequestSender,
-            DbResponse,
-        },
-    },
-    encode_args::encode_constructor_args,
-};
-
-use assertion_da_core::{
-    DaFetchResponse,
-    DaSubmission,
-    DaSubmissionResponse,
-};
+use crate::api::types::DbRequestSender;
 
 use alloy::{
-    primitives::{
-        B256,
-        Bytes,
-        keccak256,
-    },
+    primitives::Bytes,
     signers::{
         Signature,
-        Signer,
         local::PrivateKeySigner,
     },
 };
@@ -52,7 +32,6 @@ use serde_json::{
     Value,
     json,
 };
-use tokio::sync::oneshot;
 use uuid::Uuid;
 
 use http_body_util::BodyExt;
@@ -62,9 +41,7 @@ use hyper::{
 };
 use tracing::{
     debug,
-    error,
     info,
-    trace,
     warn,
 };
 
@@ -215,8 +192,8 @@ where
     let result = match method {
         #[cfg(feature = "debug_assertions")]
         "da_submit_assertion" => {
-            let rax = raw_bytecode_assertion(
-                json_rpc.clone(),
+            let res = accept_bytecode_assertion(
+                &json_rpc,
                 db,
                 signer,
                 &request_id,
@@ -232,109 +209,19 @@ where
             .record(req_start.elapsed().as_secs_f64());
             gauge!("api_requests_active", &labels).decrement(1);
 
-            rax
+            res
         }
         "da_submit_solidity_assertion" => {
-            let da_submission: DaSubmission = match serde_json::from_value(
-                json_rpc["params"][0].clone(),
-            ) {
-                Ok(da_submission) => da_submission,
-                Err(err) => {
-                    warn!(target: "json_rpc", method = "da_submit_solidity_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, error = %err, "Failed to parse DaSubmission payload");
-                    return Ok(rpc_error_with_request_id(
-                        &json_rpc,
-                        -32602,
-                        format!("Invalid params: Failed to parse payload {err:?}").as_str(),
-                        &request_id,
-                    ));
-                }
-            };
-
-            debug!(target: "json_rpc", compiler_version = da_submission.compiler_version, da_submission.solidity_source , "Compiling solidity source");
-
-            let bytecode = match compile_solidity(
-                da_submission.assertion_contract_name.as_str(),
-                da_submission.solidity_source.as_str(),
-                da_submission.compiler_version.as_str(),
-                docker,
-            )
-            .await
-            {
-                Ok(bytecode) => bytecode,
-                Err(err) => {
-                    warn!(target: "json_rpc", method = "da_submit_solidity_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, error = %err, compiler_version = da_submission.compiler_version, contract_name = da_submission.assertion_contract_name, "Solidity compilation failed");
-                    return Ok(rpc_error_with_request_id(
-                        &json_rpc,
-                        -32603,
-                        &format!("Solidity Compilation Error: {err}"),
-                        &request_id,
-                    ));
-                }
-            };
-
-            let encoded_constructor_args = match encode_constructor_args(
-                &da_submission.constructor_abi_signature,
-                da_submission.constructor_args,
-            ) {
-                Ok(encoded_args) => encoded_args,
-                Err(err) => {
-                    warn!(target: "json_rpc", method = "da_submit_solidity_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, error = %err, constructor_abi = da_submission.constructor_abi_signature, "Constructor args ABI encoding failed");
-                    return Ok(rpc_error_with_request_id(
-                        &json_rpc,
-                        -32603,
-                        &format!("Constructor args ABI Encoding Error: {err}"),
-                        &request_id,
-                    ));
-                }
-            };
-
-            let mut deployment_data = bytecode.clone();
-            deployment_data.extend_from_slice(&encoded_constructor_args);
-
-            // Hash to get ID
-            let id = keccak256(&deployment_data);
-            let prover_signature = match signer.sign_hash(&id).await {
-                Ok(sig) => sig,
-                Err(err) => {
-                    warn!(target: "json_rpc", method = "da_submit_solidity_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, error = %err, "Failed to sign assertion");
-                    return Ok(rpc_error_with_request_id(
-                        &json_rpc,
-                        -32604,
-                        "Internal Error: Failed to sign Assertion",
-                        &request_id,
-                    ));
-                }
-            };
-            trace!(target: "json_rpc", ?id, ?prover_signature, bytecode_hex = ?deployment_data, "Compiled solidity source");
-
-            let stored_assertion = StoredAssertion::new(
-                da_submission.assertion_contract_name.clone(),
-                da_submission.compiler_version.clone(),
-                da_submission.solidity_source,
-                deployment_data,
-                prover_signature,
-                da_submission.constructor_abi_signature,
-                encoded_constructor_args,
-            );
-
-            let res = process_add_assertion(
-                id,
-                stored_assertion,
-                db,
+            let res = accept_solidity_assertion(
                 &json_rpc,
-                request_id,
-                &client_ip,
+                db,
+                signer,
+                docker,
+                &request_id,
                 &json_rpc_id,
+                &client_ip,
             )
             .await;
-
-            if let Ok(ref response) = res {
-                if !response.contains("\"error\"") {
-                    info!(target: "json_rpc", method = "da_submit_solidity_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, ?id, contract_name = da_submission.assertion_contract_name, compiler_version = da_submission.compiler_version, "Successfully compiled Solidity assertion");
-                } else {
-                    warn!(target: "json_rpc", method = "da_submit_solidity_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, ?id, contract_name = da_submission.assertion_contract_name, compiler_version = da_submission.compiler_version, "Failed to process Solidity assertion");
-                }
-            }
 
             histogram!(
                 "da_request_duration_seconds",
@@ -345,45 +232,9 @@ where
             res
         }
         "da_get_assertion" => {
-            let id = match json_rpc["params"][0].as_str() {
-                Some(id) => id,
-                None => {
-                    warn!(target: "json_rpc", method = "da_get_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, "Invalid params: missing id parameter");
-                    return Ok(rpc_error_with_request_id(
-                        &json_rpc,
-                        -32602,
-                        "Invalid params: Didn't find id",
-                        &request_id,
-                    ));
-                }
-            };
-
-            // Validate hex input
-            let id: B256 = match id.trim_start_matches("0x").parse() {
-                Ok(id) => id,
-                _ => {
-                    warn!(target: "json_rpc", method = "da_get_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, id = id, "Failed to decode hex ID");
-                    return Ok(rpc_error_with_request_id(
-                        &json_rpc,
-                        -32605,
-                        "Internal Error: Failed to decode hex of id",
-                        &request_id,
-                    ));
-                }
-            };
-
-            debug!(target: "json_rpc", ?id, "Getting assertion");
-
             let res =
-                process_get_assertion(id, db, &json_rpc, request_id, &client_ip, &json_rpc_id)
-                    .await;
+                retreive_assertion(&json_rpc, db, &request_id, &json_rpc_id, &client_ip).await;
 
-            // Log success for get_assertion if not an error response
-            if let Ok(ref response) = res
-                && !response.contains("\"error\"")
-            {
-                info!(target: "json_rpc", method = "da_get_assertion", %request_id, %client_ip, json_rpc_id = %json_rpc_id, ?id, "Successfully retrieved assertion");
-            }
             histogram!(
                 "da_request_duration_seconds",
                 "method" => "get_assertion",
@@ -468,104 +319,7 @@ impl StoredAssertion {
     }
 }
 
-pub async fn process_add_assertion(
-    id: B256,
-    stored_assertion: StoredAssertion,
-    db: &DbRequestSender,
-    json_rpc: &Value,
-    request_id: Uuid,
-    client_ip: &str,
-    json_rpc_id: &Value,
-) -> Result<String> {
-    // Store in database
-    let (tx, rx) = oneshot::channel();
-
-    let ser_assertion = match bincode::serialize(&stored_assertion) {
-        Ok(ser) => ser,
-        Err(err) => {
-            error!(target: "json_rpc", %request_id, %client_ip, json_rpc_id = %json_rpc_id, error = %err, "Failed to serialize assertion for database storage");
-            return Ok(rpc_error_with_request_id(
-                json_rpc,
-                -32603,
-                "Failed to deserialize assertion internally.",
-                &request_id,
-            ));
-        }
-    };
-
-    let req = DbRequest {
-        request: DbOperation::Insert(id.to_vec(), ser_assertion),
-        response: tx,
-    };
-
-    db.send(req)?;
-
-    let result = DaSubmissionResponse {
-        id,
-        prover_signature: stored_assertion.prover_signature.as_bytes().into(),
-    };
-
-    match rx.await {
-        Ok(_) => {
-            debug!(target: "json_rpc", %request_id, %client_ip, json_rpc_id = %json_rpc_id, ?id, "Successfully stored assertion in database");
-            Ok(rpc_response(json_rpc, result))
-        }
-        Err(err) => {
-            error!(target: "json_rpc", %request_id, %client_ip, json_rpc_id = %json_rpc_id, error = %err, "Database operation failed for assertion storage");
-            Ok(rpc_error_with_request_id(
-                json_rpc,
-                -32603,
-                "Failed to write to database. Please try again later.",
-                &request_id,
-            ))
-        }
-    }
-}
-
-async fn process_get_assertion(
-    id: B256,
-    db: &DbRequestSender,
-    json_rpc: &Value,
-    request_id: Uuid,
-    client_ip: &str,
-    json_rpc_id: &Value,
-) -> Result<String> {
-    let (tx, rx) = oneshot::channel();
-    let req = DbRequest {
-        request: DbOperation::Get(id.to_vec()),
-        response: tx,
-    };
-
-    db.send(req)?;
-    let res = rx.await?;
-
-    match res {
-        Some(DbResponse::Value(val)) => {
-            let stored_assertion: StoredAssertion = bincode::deserialize(&val)?;
-
-            let result = DaFetchResponse {
-                solidity_source: stored_assertion.solidity_source,
-                bytecode: stored_assertion.bytecode.into(),
-                prover_signature: stored_assertion.prover_signature.as_bytes().into(),
-                encoded_constructor_args: stored_assertion.encoded_constructor_args,
-                constructor_abi_signature: stored_assertion.constructor_abi_signature,
-            };
-
-            Ok(rpc_response(json_rpc, result))
-        }
-        None => {
-            warn!(target: "json_rpc", %request_id, %client_ip, json_rpc_id = %json_rpc_id, ?id, "Assertion not found in database");
-            Ok(rpc_error_with_request_id(
-                json_rpc,
-                404,
-                "Assertion not found",
-                &request_id,
-            ))
-        }
-    }
-}
-
-fn rpc_response<T: Serialize>(request: &Value, result: T) -> String {
+pub fn rpc_response<T: Serialize>(request: &Value, result: T) -> String {
     json!({
         "jsonrpc": "2.0",
         "result": result,
@@ -610,13 +364,18 @@ pub fn rpc_error_with_request_id(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::api::DbRequest;
+    use crate::api::types::DbOperation;
     use crate::api::{
         db::listen_for_db,
         serve,
     };
     use alloy::primitives::hex;
+    use alloy::primitives::keccak256;
+    use alloy::signers::Signer;
     use sled::Config as DbConfig;
     use tempfile::TempDir;
+    use tokio::sync::oneshot;
     use tokio::{
         net::TcpListener,
         sync::mpsc,
